# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalDecisionRequest": {
        "description": "The body of an approve or reject.\n\nA **body**, not query parameters, and this is the point of the rewrite. The old handlers took\n`approver: str = \"admin\"` and `rejector: str = \"admin\"` in the query string, so the caller\nsupplied the identity that the audit record would attribute the decision to \u2014 and defaulted it\nto an administrator. The approver is now taken from the verified principal and cannot be\nexpressed on the wire at all; there is deliberately no `approver` field here.\n\n`comment` carries the reviewer's reason into `approvals.comment` and into the audit record's\n`reason`. `expected_version` is the change set's `version` as the client last saw it.",
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "expected_version": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "The version the client displayed. When supplied, a change set that has moved on produces 409 change-set-conflict instead of a decision on stale state.",
            "title": "Expected Version"
          }
        },
        "title": "ApprovalDecisionRequest",
        "type": "object"
      },
      "ApprovalRead": {
        "description": "A human decision, from `approvals`.\n\n`approver_id` is a `users.id`, not a name \u2014 the column is a foreign key with `ON DELETE\nRESTRICT`, so a decision cannot be attributed to a user who does not exist, and the user\nbehind a decision cannot be deleted. The old surface took the approver as a query-string\nstring, which could name anybody.",
        "properties": {
          "approver_id": {
            "format": "uuid",
            "title": "Approver Id",
            "type": "string"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "description": "One of approved, rejected (APPROVAL_STATUSES).",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "approver_id",
          "status",
          "created_at"
        ],
        "title": "ApprovalRead",
        "type": "object"
      },
...
      "ChangeItemRead": {
        "description": "One file's worth of a change set, from `change_items`.\n\n`old_content` and `new_content` are included because the diff viewer needs them and the\nagent's apply is verified against them. `old_hash`/`new_hash` are carried separately rather\nthan recomputed by the client: \u00a712.6 step 10 asserts the on-disk result against the hash the\nbackend recorded, and a client-computed hash would be checking the client against itself.",
        "properties": {
          "action": {
            "description": "One of create, update, delete (CHANGE_ITEM_ACTIONS).",
            "title": "Action",
            "type": "string"
          },
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "new_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Content"
          },
          "new_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Hash"
          },
          "old_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Old Content"
          },
          "old_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Old Hash"
          },
          "ordinal": {
            "title": "Ordinal",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "file_path",
          "action",
          "ordinal"
        ],
        "title": "ChangeItemRead",
        "type": "object"
      },
      "ChangeSetDetail": {
        "description": "A change set with the items that constitute its diff and the decisions taken on it.",
        "properties": {
          "applied_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "approvals": {
            "items": {
              "$ref": "#/components/schemas/ApprovalRead"
            },
            "title": "Approvals",
            "type": "array"
          },
          "blast_radius_score": {
            "title": "Blast Radius Score",
            "type": "integer"
          },
          "blast_radius_verdict": {
            "title": "Blast Radius Verdict",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "generation_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Run Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChangeItemRead"
            },
            "title": "Items",
            "type": "array"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "status": {
            "description": "One of \u00a73.6's thirteen states (CHANGE_SET_STATUSES).",
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "project_id",
          "status",
          "origin",
          "blast_radius_score",
          "blast_radius_verdict",
          "version",
          "created_at",
          "items",
          "approvals"
        ],
        "title": "ChangeSetDetail",
        "type": "object"
      },
      "ChangeSetPage": {
        "description": "A page of change sets plus the cursor that fetches the next one.\n\nKeyset rather than offset, matching the audit surface: an offset shifts under insertion, so a\nreviewer paging through pending change sets while new ones arrive would see rows twice or not\nat all.",
        "properties": {
          "change_sets": {
            "items": {
              "$ref": "#/components/schemas/ChangeSetSummary"
            },
            "title": "Change Sets",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "change_sets"
        ],
        "title": "ChangeSetPage",
        "type": "object"
      },
      "ChangeSetStatusFilter": {
        "enum": [
          "draft",
          "validating",
          "rejected_by_policy",
          "blocked",
          "pending_approval",
          "approved",
          "rejected",
          "expired",
          "applying",
          "applied",
          "rolled_back",
          "conflicted",
          "reverted"
        ],
        "title": "ChangeSetStatusFilter",
        "type": "string"
      },
      "ChangeSetSummary": {
        "description": "A change set without its items, for the list view.",
        "properties": {
          "applied_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "blast_radius_score": {
            "title": "Blast Radius Score",
            "type": "integer"
          },
          "blast_radius_verdict": {
            "title": "Blast Radius Verdict",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "generation_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Run Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "status": {
            "description": "One of \u00a73.6's thirteen states (CHANGE_SET_STATUSES).",
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "project_id",
          "status",
          "origin",
          "blast_radius_score",
          "blast_radius_verdict",
          "version",
          "created_at"
        ],
        "title": "ChangeSetSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ForgeOps API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/approvals": {
      "get": {
        "description": "A keyset page of the caller's tenant's change sets, newest first.",
        "operationId": "list_change_sets_api_v1_approvals_get",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ChangeSetStatusFilter"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeSetPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List change sets awaiting or past decision",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/v1/approvals/{change_set_id}": {
      "get": {
        "description": "One change set with its `change_items` and every decision recorded against it.",
        "operationId": "get_change_set_api_v1_approvals__change_set_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "change_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Change Set Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeSetDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read one change set and its diff",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/v1/approvals/{change_set_id}/approve": {
      "post": {
        "description": "Approve, then mint authority and deliver the envelope \u2014 the full \u00a711.6 transit.\n\nThe approver is `principal.user_id` and cannot be supplied by the caller. Delegated to the\nchokepoint rather than reimplemented, so approval keeps its optimistic-concurrency check, its\n`approvals` row, its audit event and its re-evaluation of policy at apply time.",
        "operationId": "approve_change_set_api_v1_approvals__change_set_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "change_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Change Set Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approve Change Set Api V1 Approvals  Change Set Id  Approve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve a pending change set",
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/api/v1/approvals/{change_set_id}/reject": {
      "post": {
        "description": "Refuse a pending change set: \u00a73.6's `pending_approval \u2192 rejected` edge.\n\nNothing is minted and nothing is delivered, because there is nothing to apply. The refusal is\nstill a governance event with a row and an audit record \u2014 under the old dictionary it was\nneither.",
        "operationId": "reject_change_set_api_v1_approvals__change_set_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "change_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Change Set Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reject Change Set Api V1 Approvals  Change Set Id  Reject Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject a pending change set",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/v1/approvals/{change_set_id}/revert": {
      "post": {
        "description": "Compile and authorise the reverse change set (\u00a711.6, D-66).\n\nNamed `revert`, not `rollback`, because \u00a73.6's edge out of `applied` is `applied \u2192 reverted` and\nthe old handler's name matched no edge in the state machine. `rolled_back` is a different state\nreached from `applying` when an apply fails \u2014 conflating them would make the two\nindistinguishable on the record.\n\nA revert is a mutation and goes through all six stages with its own fresh authority: reusing the\noriginal's would make rollback a privileged back door.",
        "operationId": "revert_change_set_api_v1_approvals__change_set_id__revert_post",
        "parameters": [
          {
            "in": "path",
            "name": "change_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Change Set Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Revert Change Set Api V1 Approvals  Change Set Id  Revert Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revert an applied change set",
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/api/v1/mcp/apps/{name}": {
      "get": {
        "description": "MCP App descriptor. The host page is served separately at /apps/{name}/host.",
        "operationId": "get_app_descriptor_api_v1_mcp_apps__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get App Descriptor",
        "tags": [
          "mcp"
        ]
      }
    },
...
  }
}